Update to include support for cookies
Add support for wordpress sites that require a particular cookie(s) in
order to be accessible. For example, this could happen on a company
intranet where the SSO mechanism does not support basic authorization.

Update client to accept an additional cookie value and pass that along
to the underlying connection when it is set.

This fixes issue #36.