Fix #37 - Allow custom prefixes on method names
This allows bare method names to be passed to Client#execute and be
prefixed by "wp." as before, but allows method names with a namespace
prefix to pass through unmodified. This is determined by whether or not
there is a '.' present in the method name.

This implementation does mean that custom methods must have some dotted
prefix, since any bare method names are assumed to need "wp.", but this
seems to be a reasonable requirement, matching API convention.